moved tests into new tests folder
