+++ artscii/src/tests/App.test.js
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
-import App from "./App";
+import App from "../app/App";
 
 test("renders app component", () => {
   render(
