+++ artscii/src/tests/testValidateDisplayManger.test.js
@@ -1,0 +1,16 @@
+// import { validModes } from './validateDisplayManagerProps';
+// import { DisplayManager } from './DisplayManager';
+// import { expect, jest } from '@jest/globals';
+// const fs = require('fs');
+// const path = require("path");
+
+// const testImg = fs.readFileSync(path.resolve(__dirname, '../../assets/fry.png'));
+// let testProps;
+
+// beforeEach(() => {
+//     testProps = {displayMode: 'ascii', src: testImg}
+// })
+
+// const validateDisplayManagerProps = jest.fs();
+
+// // Tests
